Add initial DXE Core
DXE is an execution phase defined in the Platform Initialization (PI)
Specification maintained by the UEFI working group. A "DXE Core" is
the "core" component responsible for initialization, management, and
execution of the DXE phase. This includes high-level responsibilities
such as:

- CPU and hardware initialization and configuration
- Dispatch and dependency services
- Memory allocation and resource management
- Debug facilities to understand and debug DXE code execution
- Runtime services used to execute code at OS runtime
- Event and protocol services and management
- Support for performance related services and functionality
- Support for synchronization of hardware resources
- And more...

This initial DXE Core has support for all of these areas with some
in the current state being developed further before a major release
of the repo content is planned. The items planned for this release
are tracked in the repository as GitHub issues.

For further information see the accompnaying code, documentation, and
tests.

Co-authored-by: John Schock <joschock@microsoft.com>
Co-authored-by: Michael Kubacki <michael.kubacki@microsoft.com>
Co-authored-by: Joey Vagedes <joeyvagedes@microsoft.com>
Co-authored-by: Mathieu Gravel <magravel@microsoft.com>
Co-authored-by: Oliver Smith-Denny <osde@microsoft.com>
Co-authored-by: Chris Fernald <chfernal@microsoft.com>
Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>