dxe_core: Update toolchain to 1.80.0
Updates the toolchain to 1.80.0, which resulted in two distinct new
panics during testing due to updates to the
`core::slice::from_raw_parts` which now verifies specific conditions and
panics if those are not met.

1. The first was due a null pointer for the dxe_core image address for
tests in image.rs. For testing, we generated a hob list with a null ptr
for the dxe_core_image address to satisfy `install_dxe_core_image`. With
the new changes, a check had to be added for a null ptr to avoid a panic
in `core::slice::from_raw_parts`.

2. The second was due an invalid (but not null) pointer for a test in
the fv parsing code. Previously, an invalid address was passed to the fv
parsing code to ensure we would catch the invalid address and return.
With the changes to `core::slice::from_raw_parts`, the code panic'd, so
a check to validate the fv_header signature was added before the call to
`core::slice::from_raw_parts` instead of after, to avoid the panic. This
change occurred in mu_rust_pi, thus the update to 5.0.1 and the update
to uefi-core.