Makefile: Add build-aarch64, build-x64
At some point in time, build-arch evolved into building the dxe_core
Qemu efi images when it was suppose to be building all crates with the
specific uefi target triple.

This commit adds build-aarch64 and build-x64 commands to the Makefile
to build all crates with the specific uefi target triple as expected,
and matches the format in uefi-core. building the qemu binaries has
been renamed to build-bin, however this will eventually be dropped when
the binaries move to their own location outside the crate.