Bugfix: Fix broken headers in mdbook (#104)
## Description

A previous commit to satisfy markdownlint requirements broke the headers
in the mdbook. The `#` header shows up in the mdbook as a new section
outside of chapters. Changing to `##` ended up breaking the intended
formatting of the mdbook as those headers no longer showed up.

| Broken | Intended |
|--------|----------|
|
![image](https://github.com/user-attachments/assets/44ca5b34-e174-4d15-93b3-86a016e7cbff)
|
![image](https://github.com/user-attachments/assets/7b5c16ca-618c-4783-8535-792c9d16a414)
|

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Locally ran mdbook and verified it looks as expected

## Integration Instructions

N/A