Validate Documentation
Adds a validation step to the CI workflow to ensure that the
documentation will build with no warnings or errors. Relies on a new
cargo make command `doc` to build the documentation. This allows
repositories to customize the build process for documentation (including
adding features). Adds `-D warnings` to ensure that warnings are treated
as errors.