Relax crate dependency versions
This commit relaxes the crate dependency versions in the Cargo.toml file
to allow for the dependency resolver to more easily find common crate
versions, which can reduce the size of the final binary by not needing
to use multiple versions of the same crate in the final consumer.

This also reduces the likely hood of a build error in which the same
trait is compiled using two different versions, and due to name
mangling, are detected as two different traits, causing a build error.

Due to the relaxed versions, the CI system will now upload the lock file
that is generated, so that we have reproducible builds when necessary.