.github: Add release workflow
This commit adds a reusable workflow that is called manually by a
maintainer that will publish all crates in the workspace to crates.io.

The version used is the current tag of the release draft in the same
repository. If there are multiple (or no) drafts, the workflow will
abort. The steps are:

1. Test that we can compile and publish with a dry run
2. Create a commit updating all versions to the tag
3. Publish to crates.io based off that new commit
4. Push the commit to the branch
5. Wait for release updater to finish
6. Publish the release on github