dxe_core: Improve DxeCore public interface
This commit improves the DxeCore public interface in many ways:

1. Splits the core into two distinct parts: pre-init and post-init
mainly for controlling the types of configuration you can do (no
allocations pre-init and allocations post-init).

2. The interface for the post-init DxeCore is locked down and can only
be accessed by calling `initialize` on the pre-init DxeCore.

3. Improves the interface by no longer requiring drivers to be static.
This is accomplished by holding the driver in a Box and leaking it,
effectively making it static, which is required by corosensei. This is
done in a safe way as we consume the drivers, thus no driver can be
leaked more than once.

4. Sets the example on how to do configuration for the DXE Core and for
the drivers. That is to say that `default` must be implemented, then you
can override the default configurations by calling the `with_<config>`
methods.