.github: Add release workflow (#113)
## Description

This commit adds a reusable workflow that is called manually by a
maintainer that will publish all crates in the workspace to crates.io.

The version used is the current tag of the release draft in the same
repository. If there are multiple (or no) drafts, the workflow will
abort. The steps are:

1. Test that we can compile and publish with a dry run
2. Create a commit updating all versions to the tag
3. Publish to crates.io based off that new commit
4. Push the commit to the branch
5. Wait for release updater to finish
6. Publish the release on github

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated full process works as expected through running the workflow on
a fork. The testing run did not login to cargo, and skipped the publish
step, but all other things work as expected including the dry run, tag
detected, and release publishing.

## Integration Instructions

N/A