.github: reconvene matrix jobs
The ci workflow generates a matrix of jobs to execute and github cannot
specify those jobs as needing to complete before a PR can be merged. To
get around this, we hardcode one final job that will be run after all
the other jobs are done. This will allow branch rules to be set
requiring that the final job is successful before a PR can be merged.