dxe_core: Add EFI_MEMORY_MAP trace logging
This adds a new target for logging: efi_memory_map.
When set to info level for a platform, the EFI_MEMORY_MAP
will be pretty printed in the log whenver it is created.
Info logging was chosen over trace because trace logging
prints the file name before each line, which takes up
too much room for what this table needs to be easily
read and it is distracting with so many lines.

This commits also changes the MAT logic to use the
memory descriptor pretty print logic created in this
commit, though it does not tie it to the efi_memory_map
logging target or any other, as the MAT is created less
often and has fewer members and is useful to be in the
log by default.

The column sizes and individual column member spacing was
tested and adjusted so it lined up visually, even though
the numbers in the format strings appear off.