allocator: Add ensure_capacity function
This commit adds a new ensure_capacity API to the allocator
that checks if a given size and alignment are free in the
given allocator. If not, it calls expand to allocate more
pages in the pool.

This API is intended for dxe_core only when it needs to
pre-allocate memory in memory sensitive operations, such as
constructing the EFI_MEMORY_MAP or allocating page table
memory.

By ensuring the pool allocator has enough memory for its
allocations, it can ensure that recursive calls to the GCD
will not be made.

The get_memory_map code is updated here also, not to use
the new API as was the original design point, but to
pre-allocate the merged descriptor vector. Same goal,
but different implementation.

This API will be consumed when mu-paging is integrated
into dxe_core.