Add uefi_cpu_init and uefi_interrupt crates
With PR #133 adding CPU Architecture specific default trait
implementations for some DXE Core traits, they have now become
implementation details of the DXE Core rather than platform customizable
traits. Due to this, these traits (uefi_cpu_init and uefi_interrupt) now
make more sense in the uefi-dxe-core repo rather than the uefi-core
repo, which is meant to be a place for reference implementations of
platform customizable traits for the DXE core.

The trait continues to provide a point of abstraction for the DXE Core,
but is now a point of abstraction for future CPU architecture specific
implementations rather than platform specific implementations. A
platform, if necessary, can still provide its own implementation of the
trait, but the expectation is that the default implementation will be
sufficient for most platforms.