uefi_test: cleanup macro usage
in the uefi_test macro, we originally manually expanded the code
generated by the `distributed_slice` macro from linkme, as we were under
the impression that the path to the linkme crate was hardcoded as
`::linkme`. After additional review of the linkme code (because it was
not mentioned in the linkme documentation), it was discovered that a
attribute macro is available to override the path to the linkme crate
(`#[linkme(crate = "path::to::linkme")]`).

This commit removes the manual expansion of the macro and replaces it
with the macro itself, using the path override attribute to specify the
path to the linkme crate.