image loader: Fix Freeing Stack
Currently image stacks can't be freed because the memory code
cannot apply attributes across blocks, so core_free_pages
returns not found. There are multiple blocks because of the
stack guard having a different attribute (RP) than the rest of
the stack (XP). This is fixed by removing the RP attribute and
setting XP before freeing the stack, so that the blocks are
coalesced and core_free_pages can free the single block.

There is a larger discussion on attribute handling in the memory
subsystem that will be revisted later.