fv.rs: Fix FV Protocol ReadFile() behavior
The EFI_FIRMWARE_VOLUME2_PROTOCOL.ReadFile() definition in the PI
Specification states:

" Pointer to a pointer to a buffer in which the file contents are
returned, not including the file header. See 'Description' below
for more details on the use of the Buffer parameter."

However, the implementation currently returns the header followed
by the contents, in turn also returning the header + content size
instead of just the file content size.

This change updates read_file() to match the PI Specification
described behavior and match the C code behavior.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>