protocol_db: remove PROTOCOL_DB statics
This removes the usage of PROTOCOL_DB statics in the protocol related
boot services functionality. This enables easier unit testing of these
functionalities, as you do not have to be concerned with locking and
configuring the static PROTOCOL_DB.

All other usage of the PROTOCOL_DB static is still in place, but are
managed in the boot_services static rather than being a static itself.
As more content is moved, less and less of the PROTOCOL_DB static will
be used, and more content can be easily tested.