protocol_db: move the protocol_db crate into the dxe_core
Moves the protocol_db crate into the dxe_core as the protocol_db module.

Due to this move, and the fact that uefi_depex will not be moved into
the dxe_core, the dependency of uefi_depex on uefi_protocol_db is
removed by updating uefi_depex to accept a list of guids for it's
evaluation rather than the SpinLockedProtocolDb struct.
SpinLockedProtocolDb is updated to produce a vector of guids (Copied)
that are registered with the SpinLockedProtocolDb.

Additionally removes all doc tests from protocol_db as they only work on
public modules and by moving it into the dxe_core, it is no longer
public.