rust-tool-cache: use cargo binstall
cargo install has shown to be an unreliable way to compile and install
tools for any project that does not use the current latest version of
rust. This is because some tools allow for relaxed versions of their
dependencies, which will cause cargo to attempt to use the latest
version of the dependency, which may not be compatible with the current
version of rust being used.

This change switches to using cargo binstall, which simply downloads
the compiled tool, with a fallback to cargo install if the tool does
not support binstall.