Relocate the HobList to Allocated Memory
PEI does not create a memory allocation HOB for the HOB list
itself. When DXE Core receives the "contract" is that the HOB
list lives in unallocated memory until enough resource descriptor
HOBs have been added to relocate the HOB list in an area of
memory that can be allocated for it.

This change is required for integrating paging because if the
HOB list continues to live in unallocated memory, accesses to it will
fault when we install the new page table. It worked with the page table
from PEI because everything is mapped RWX there and the HOB list is not
located in memory that can be allocated by the allocators.