Update set_memory_space_attributes to Split Attributes Across Descriptors
Currently, set_memory_space_attributes does not allow for attributes to
be set across multiple descriptors. However, the PI spec allows for this
through this API and the CPU arch protocol and memory attribute protocol
allow for this and are thin wrappers over the set_memory_space_attributes
function.

This fixes an issue with Windows booting on a failure path: when Windows sees
that previous boots have shutdown unexpectedly, it will boot to WinRE. When
doing so, it unloads winload and sets the entire image space to clear attributes.
This fails in the Rust core, because the winload image sections have different
attributes set, meaning that they have different GCD descriptors and so
set_memory_space_attributes disallows setting the attributes.

This commit simply checks at the top level set_memory_space_attributes code
to see if the range splits multiple descriptors and calls the inner function
once for each descriptor (or part of descriptor) in the range.