release workflow update to run on github release
Updates the release workflow to update on a github repository release
event. As before, this workflow will publish all crates with the
specified version number. The main difference is that instead of pushing
a commit directly to the default branch of the repository, it will
instead create a new branch and open a pull request.