Align EFI_MEMORY_ATTRIBUTES_PROTOCOL to UEFI Spec
The memory attributes protocol implementation in uefi-dxe-core
was taking some liberties with its implementation w.r.t. the
UEFI spec, namely it was returning some error codes not
supported by the spec and more importantly, it had a bug in that
it relied on lower layers of code to split the request per GCD
descriptor, but the protocol code took attributes from the first
descriptor of the range and attempted to apply that to the entire
range (as these APIs are different than the GCD APIs that are used
under the hood and so need some massaging).

This commit rectifies the bug and aligns the protocol API to the
spec.

This fixes a bug seen when bootmgr attempts to unload winload while
attempting automatic recovery, leading to a crash due to inconsistent
GCD attributes across a range described by one GCD descriptor.