Fix DXE Core Loaded Image Protocol
DXE Core's loaded image protocol today is constructed using
PrivateImageData::new(). However, this function expects that
we are about to load this image, so it allocates an image
buffer for the image to be loaded into and sets some fields
with these assumptions.

DXE Core is unique in creating a loaded image protocol as it is
already loaded, so PrivateImageData::new() allocates a large chunk
of wasted memory for a buffer that never gets used and incorrectly
lists where DXE core is located.

This also creates a problem in PrmLoaderDxe when it searches through
loaded image protocols looking for PRM modules and crashes trying to
read the invalid DXE core base address in the loaded image protocol.

With this change, that crash is fixed and the wasted memory is not
allocated.