Update cspell config
A lot of words we're using are in some other language dictionaries,
especially the `cpp` dictionary, so use those to reduce false
positives. The words that we added to the `words` list already in
those dictionaries are removed from our `words` list to simplify
the file and reduce maintenance.

In some rare cases, hexadecimal digits are being interpreted as words
so exclude those in `ignoreRegExpList`.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>