.git-blame-ignore-revs: Update commit sha.
Apparently even on a rebase & ff when both commits are top of tree, the
commit sha is changed, so the second commit in the rebase that added
the .git-blame-ignore-revs file was ultimately not the correct commit
sha. This commit updates to the correct sha now that the commit has
been merged.