CI: dry run publishing to a registry
Changes to a crate's Cargo.toml package
information can result in a crate not being able
to publish to a registry (such as a description
missing). This PR adds a new step to our CI that
attempts to publish the crate to the registry
before the PR is merged. This will ensure the
break will be caught before it is merged in,
instead of during the release process.