spin_locked_gcd: Set Paging Attributes Per GCD Descriptor
When setting memory attributes, the GCD sets all GCD attributes
and then moves to do the paging attributes as a whole. This works
when the entire GCD operation succeeded, because that has guaranteed
the GCD has a consistent view of that range of memory, however, in
the case of failure this can create a torn state where the GCD is
partially updated and the page table is not updated.

To avoid this torn state, the page table is updated after every
GCD descriptor update.