Add Code Organization doc
The Rust DXE Core is a complex system with many constituent parts.
This document describes the organization of the overall codebase,
including the key dependencies that are shared between the Rust DXE
Core and other components. The goal is to provide a high-level
overview of these relationships.

This is meant to be a living document, and as the code base
evolves, this document should be updated to reflect the current
state.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>