Add Compatibility Mode Support
This patch adds compatibility mode (to be documented soon). This
adds a build flag (disabled by default) to enable compatibility
mode when launching EFI_APPLICATIONS without the NX_COMPAT flag.
The primary use case for this is to launch shim/grub which do not
work with memory protections.

When compatibility mode is enabled the following occurs:
- The loaded image that triggered the mode has its memory set to RWX
- All newly allocated memory is allocated RWX
- All existing EFI loader code and data memory is marked RWX
- The memory attributes protocol is uninstalled
- The range 0 - 0xA000 is set to RWX if it is system memory