Temporarily disable setting read-only sections
Commit 30fa804 made sections in a PE/COFF image that are not writable
but readable to have the read-only attribute set. While this has
successfully booted to Windows on QEMU and no issues have been
encountered in UEFI images, it does fail on physical Intel platforms,
at least in some cases.

That is being tracked in:
https://github.com/OpenDevicePartnership/uefi-dxe-core/issues/365

To allow physical platforms to release the code is temporarily
commented. It is left in the codebase as a reminder that it should
be followed up on after the issue is closed.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>