Switch to a public registry (`patina-fw`)
Closes #398

Closes #398

Moves to a temporary Azure Artifacts feed that serves as the Patina crate registry until the project's crates are published to crates.io.

Connect to the crate in a `.cargo/config.toml` with:

```
[registries]
temporary = { index = "sparse+https://pkgs.dev.azure.com/patina-fw/artifacts/_packaging/temporary/Cargo/index/" }
```

Feed URL:

https://dev.azure.com/patina-fw/artifacts/_artifacts/feed/temporary

Additional Setup Instructions:

https://dev.azure.com/patina-fw/artifacts/_artifacts/feed/temporary/connect

---

To clarify terminology the following are involved in the public feed:

- Organization: `patina-fw`
- Project: `artifacts`
- Feed: `temporary`

Originally, I tried making "crates.io" an upstream of the `temporary` feed so we could use a single feed for our published crates while also caching upstream crates using contributor permission in CI and for local users.

This would only work if the `artifacts` repository was set to `enterprise` access level, not `public`. In that case, the feed is also not public, which defeats the purpose of this change. A `401` occurs when the feed attempts to pull in the crate version from crates.io. This happens whether using a PAT (with proper permissions) or MSAL with `msrustup` as an authenticator. By simply making the project, `enterprise` both approaches begin working with the exact same token.

Therefore, the PR has been updated to pull all crates on crates.io from crates.io and only allow publishing to the `patina-fw` feed and also set up all crates published to that feed to be fetched from there.

---

`cargo release` does not seem to honor the `registry` setting for each dependency in the workspace Cargo.toml file. It was working previously based on the complete replacement of `crates-io` with:

```toml
[source.crates-io]
replace-with = "UefiRust"
```

Due to *Special Note* above, that's not possible with this registry as we can only use it for crates not being pulled from crates.io upstream.

After trying several approaches, the one that seems to work and is the least hacky is to patch each crate that would go to the `crates-io` registry to the local crate with:

```toml
[patch.crates-io]
dxe_core = { path = "core/dxe_core" }
```

Therefore, an update has been made to do that for every crate published in this repo.

An outstanding issue is that the `mtrr` and `paging` crates are not local to this repo but also not on crates.io. Therefore, this does not directly apply to those crates.

There are at least three way to solve that:

1. Make the `mtrr` and `paging` crates public and publish their crates to crates.io
2. Move the `mtrr` and `paging` crates to the `patina` repo so they become "local paths"
3. Patch the `mtrr` and `paging` crates as git dependencies to their current ODP repo locations

- (1) is ideal and simplest if we're ready to publish those crates
  - Note: `paging` is already taken on crates.io, so it will need to be renamed, which is inevitable anyway.
- (2) is less ideal since it defeats the separation of these crates into their own repos and the separation, we wanted for those crates from the rest of the Patina code.
- (3) is even less ideal because:
  - Private ODP GitHub repos (which these are) would still need to be authenticated in CI which is another layer of complexity to add.
  - The publish step will be referencing the `main` branches of those repos which is different from the other cargo commands operating on the workspace dependencies
  - Because other cargo commands are using the "real" `mtrr` and `paging` crates, this warning will appear with the patches present when running commands like `cargo update:

    ```
    Updating `patina-fw` index
    warning: Patch `mtrr v0.1.1 ([https://github.com/OpenDevicePartnership/mtrr.git?branch=main#ab5369bc)`](https://github.com/OpenDevicePartnership/mtrr.git?branch=main#ab5369bc)%60 was not used in the crate graph.
    Perhaps you misspelled the source URL being patched.
    Possible URLs for `[patch.<URL>]`:
        patina-fw
    warning: Patch `paging v4.0.2 ([https://github.com/OpenDevicePartnership/paging.git?branch=main#218277d8)`](https://github.com/OpenDevicePartnership/paging.git?branch=main#218277d8)%60) was not used in the crate graph.
    Perhaps you misspelled the source URL being patched.
    Possible URLs for `[patch.<URL>]`:
        patina-fw
    ```

---

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [x] Includes documentation?

- `cargo update` locally

- The crate will now be published to a public ADO registry. Connect to it as follows:

```
[registries]
temporary = { index = "sparse+https://pkgs.dev.azure.com/patina-fw/artifacts/_packaging/temporary/Cargo/index/" }
```

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.co