service: IntoService Macro additional implementations
This commit updates the `IntoService` macro to implement the
`IntoService` trait for not just the concrete struct (MyStruct), but
also for a static reference to the struct (&'static MyStruct). This is
useful for cases where a service provider must have a static instance
of the struct, but still wishes to register it as a service.