Prevent double-queueing of events
This change guarantees that event notifications will not be double-queued
in order to match the Tianocore implementation of events. Before this
change, the implementation of Ord on TaggedEventNotification prevents an
event notification being queued twice if signal_event is called
one-after-the-other, but allows double-queuing in the scenario where an
event is signaled, the pending events iterator is collected, then the event
is signaled again before the event's callback is called. This scenario will
be run into when an event is signaled, then the callback for an event
earlier in the iterator signals the same event.

Closes #374