dxe_core: events: handle all event_notify events (#212)
When a TPL is restored, any notify events are processed, which may result
in a notify callback being called. This callback may result in a notify
events being added to the event queue. This commit will continuously
process all events in the event queue until the queue is empty.

Additionally, this commit moves the EVENT_NOTIFIES_IN_PROGRESS static
inside of the restore_tpl function, as it is only used in that function.

Closes #199