patina_sdk:memory: Implement mock MemoryManager
Implements a working MemoryManager implementation for the MemoryManager
trait by using the global allocator. This should be used when a mock is
needed for testing purposes, that can allocate and deallocate memory
as expected.

This commit also adds mockall's automock to the MemoryManager trait, to
simplify mocking MemoryManager interface method failures in tests.