Disable cargo release dry run - To pass CI for PR#527
## Description

Disable Cargo Release Dry Run – To Pass CI for
[[PR#425](https://github.com/OpenDevicePartnership/patina/pull/527)](https://github.com/OpenDevicePartnership/patina/pull/527)

**Why is this one-off exception needed?**
With
[[PR#425](https://github.com/OpenDevicePartnership/patina/pull/527)](https://github.com/OpenDevicePartnership/patina/pull/527),
we introduced a new feature flag(`patina_tests`) in both patina_sdk
and patina_sdk_macro crates. One relying on the other and cargo release
dry run is unable to locate the package in the registry with that
feature flag.

This change disables the dry run to allow the CI to pass for the above
PR. Once the PR is merged, we will proceed with the usual process of
generating a draft release, and the `publishrelease` workflow will
handle the actual publication of the crates.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

NA

## Integration Instructions

NA