AdvLogger: Fix UB from casting header values to atomic.
The initial implementation tried to stay as accurate to a C definition of the structures
as possible and so used u32 in cases where valuers are only atomically updated. This required
that those types be converted to atomic (internally mutable) values later which can cause undefined
behavior. This change is to make the structure itself contain the atomic versions of the types.