cspell: Remove Overly Permissive Dictionaries
This commit removes several dictionaries from cspell and
ensures that the list of allowed words is the needed list.

The software terms dictionary, in particular, was problematic,
because it contained many double wildcards, e.g. *Core*, so anything
with the substring Core in it would pass spell check, even if the
word was misspelled.

Removing the extra dictionaries also speeds up the cspell execution,
in theory, particularly without all of the wildcards because cspell
must do less searching and pattern matching to determine if words
are misspelled.

The java and python dictionaries were removed because patina is not
implemented in those languages. The rust dictionary is kept because it
has no wildcards or explicitly allowed misspellings, instead it contains
useful Rust keywords.

A few choices were made across the code base to align spellings, e.g.
re-entrancy and reentrancy were both present, so reentrancy was
aligned on.