patina_debugger: Remove monitor buffer and add simple buffering
With the newest Windbgx, the configuration for UEFI has been updated such
that the GDB extension will wait until a monitor command is completed with
an OK packet to signal that all responses have been sent. This means that buffering
the entire response into a single buffer is no longer required and the message
may be sent over multiple GDB packets. So the monitor buffer that is pre-allocated
is no longer required.

Additionally, this change introduces a `MonitorBuffer` structure that will be inserted
on the stack. This temporary stack is used to reduce the number of packets that will
be sent. Format commands in rust will result in a large number of small writes as it
formats the various types of potentially character by character. This new stack buffer
serves as an optimization to reduce those packets.