patina_debugger: Ignore null bytes read from the transport
During some operations. The debugger will read a null (0x00) bytes from
the transport. This packet is noe expected and does not appear to be related
to any input command. Generally GDB is a ASCII based protocol and does not
expected NULL characters. This change ignores such bytes and prints
and error to avoid debugger crashes.