patina_dxe_core: Convert allocators to have static lifetime
All allocators by their very nature should have a static lifetime. This is
critical for the MemoryManager API that explicitly requires this, but also
for the creation of any smart pointers like a Box, the allocator associated
with it must be static or else it may have a shorter lifetime than that of
the Box.