patina_dxe_core: Only Unmap Page 0 If It Exists
This patch updates the logic to unmap page 0 after resource
HOBs are added. It previously attempted to always do it, but
on some platforms page 0 does not exist (in which case the page
table will not map it anyway). This checks if the memory exists
before attempting to unmap, so it does not cause a debug_assert.