Remove the exclusion of patina_sdk_macro from code coverage
For some reason, the patina_sdk_macro crate keeps being added as an exclusion in the Makefile because local developers are on an older version of cargo tarpaulin then they should be, and add it as an exclusion instead of upgrading cargo tarpaulin to what is specified in the toolchain file. This commit adds back patina_sdk_macro crate so we get code coverage statistics for it.