allocator: Remove static memory type info table
This commit removes the static memory type info table. Due to the need
to keep all memory type information in a single table (as it is
published as a system table), this commit opts to give ownership of
the entire table to the GCD. When the GCD creates a new allocator, it
passes a pointer to the memory type info for that allocator, to the
allocator. This allows the allocator to self update it's memory type
information as needed, while still keeping the entire table together.

For dynamically generated allocators (e.g. the allocators not a part
of the 17 tracked memory types), a new MemoryTypeInfo is created and
leaked, for them to use.