allocator: Don't Assert if Page 0 Has Memory Allocation HOB
edk2/Mu made a change to drop DXE IPL from creating memory
allocation HOBs for page 0, which Patina was relying on.
However, this is a recent change and many existing platforms
still create the memory allocation HOBs. We should not
debug_assert in this case, but simply let the allocation fail.
We can still unmap the page for null pointer detection.

This also adjusts some stale comments.