dxe_core: Point to Reallocated Table in Dbg Image Info
Currently, the debug image info table updates the Rust slice
pointer to a newly reallocated table, but does not update the
UEFI defined structure to point to the new table, as a result
the old table gets freed and the table in memory (discovered by
the debugger) points to garbage.

This correctly points the in memory UEFI spec defined structure
to point to the new table.