patina_debugger: Implement memory poke test
The debugger will attempt to check the page tables to determine
if memory is safe to access, but this is not 100% affective as
memory will often be mapped that is not actually real or accessible.
This can happen in the inherited page tables or for MMIO regions
that may not be present.

The poke test is an architecture specific function that will set
a marker and touch the memory. During the exception entry if the
marker is set then it can be concluded that the memory is bad and
we can jump past the faulting instruction and return. This prevents
nested faults when actually accessing the memory.