allocator: Support EfiUnusableMemory memory type
Per the UEFI specification, the EfiUnusableMemory memory type is a that
specifies that the memory contains errors and is not to be used. The
allocation APIs specifically allow for this memory type to be allocated
to (1) reserve and prevent the memory from being used by the system and
(2) allow the memory to be descrived in the memory map.

This commit allows allocations to be made with the EfiUnusableMemory
memory type for this reason, but it is still considered an error to
allocate and actually use memory with this memory type. This change
allows for the BootServicesTest::MemoryAllocationServicesTest0 to pass.