patina_dxe_core: GCD: Only Map Page 0 If Existent in Compat Mode
On some ARM64 platforms, page 0 is not a real address. This is
anticipated by the GCD code and it intends to not map it in this
case. However, the GCD envisions that page 0 will not exist in
the GCD if this is the case, but on some platforms it will show up
as NonExistent (as opposed to no descriptor).

This fixes the logic to not attempt to map page 0 if it does not
exist in the GCD or if the type is non-existent.