Fix end of BootServices not reported. (#684)
## Description

Before the end of exit boot services was not reported. This was caused
by some status code not executed. After some investigation, I found that
the status code listed was unregister in an event run at the end of boot
services. This causes status code to set the end of exit boot services
to not be called.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

A perf record was generated.

## Integration Instructions

N/A