Patina: Remove Incorrect asm! Macro Options
The sti and cli options explicit set/clear the
interrupt flags in the RFLAGS register.

This is technically okay to use preserves_flags,
per the Rust documentation:
https://doc.rust-lang.org/reference/inline-assembly.html#r-asm.rules.preserved-registers,
as the interrupt flag is not one of the flags the
docs say must be preserved manually is preserves_flags
is set.

However, this may cause confusion or be changed in the future,
so this commit removes the preserves_flags option from these
as it is safer to not use that option when any instruction
changes RFLAGS.