patina_dxe_core/pecoff: Verify PointerToRawData is resolved properly
Verifies that the behavior in the following goblin fix is followed:

https://github.com/m4b/goblin/commit/04d69ad5713d8a21aa295ef4ea047524a7edd615

This verifies that the filename in the debug section is found using
an image built with 512 byte file alignment and one built with 32
byte file alignment. Both have 4k-section alignment. This verifies
that binaries with very small file alignments have PointerToRawData
calculated to the correct position in the image.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>