Remove mutable static usage from perf (#721)
## Description

Removes the mutable static usage in the performance functionality. The
original implementation needed mutable access to the static to do a
one-time initialization. This is exactly what the `OnceCell` struct is
used for, so the logic was switched to that. Some additional
synchronization logic using at atomic Boolean needed to be added to
truly support the static being `Send + Sync`.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI passes, and a successful boot to shell on Q35.

## Integration Instructions

N/A