dxe_core: Change ALLOCATE_MAX_ADDRESS to still use TopDown allocations
For consistency this commit changes MAX_ADDRESS allocations to also use
TopDown but with the max address set. This ensures that low memory is reserved
for when it is actually needed and prevents fragmenting memory.

Additionally, this change removed a condition in the allocate_pages_should_allocate_pages
test as it incorrectly assumes MAX_ADDRESS follows a bottom up scheme.