remove FunctionComponent implementation
This commit removes the `FunctionComponent` struct, associated blanket
`IntoComponent` implementation for functions, tests, and examples.
Additionally, it updates all relevant documentation to reflect this
change. Finally, it updates the `ParamFunction` trait error diagnostic
messaging provide clarify in the error message when the function used
as the entry point does not match the expected signature.