ParamFunction: Support Additional Self Types
This commit enhances the blanket `ParamFunction` trait implementation
to support functions that take `&Self::In` and `&mut Self::In` in
addition to `Self::In`.

What this translates to is that struct / enum components using the
`#[derive(IntoComponent)]` macro can now define entry point functions
that take `&self` or `&mut self` as their first parameter (in addition
to `self`). This enhancement paves the way for future support of
multi-callable dependency injectable types, including but not limited to
components.

Additionally, this commit introduces a custom error message when a
function is being used as a `ParamFunction` but does not meet the
function signature requirements. This should help users quickly identify
and rectify issues with their function signatures.