Update mdbook landing page to be clear and concise
The landing page is updated to provide a welcome page and minimal
information regarding what "Patina" is, and it's goals. It then links
across other pages in this book to provide more detailed information on
the different topics this book covers.

Specifically it also relocates the RFC Lifecycle and rust language
information to other parts of the book, as to not clutter the landing
page.