Remove "Configuration in Code" Best practices page
Remove the "Configuration in Code" best practices page as it is outdated
with the new (since written) configuration system for patina components.
The page being removed refers to what is now considered
"private configuration" (configuration specific to a component
instance).

Documentation regarding public (shared) configuration and private
configuration is located in the Patina Component Model.