.github/action.yml: Fix
The cache key is getting truncated due to double template expansion
so it looks like this in an action log:

```
Run actions/cache@v4
  with:
    path: ~/.cargo/bin/
  ~/.rustup/toolchains/

    key: Linux-rust-tools-
    enableCrossOsArchive: false
    fail-on-cache-miss: false
    lookup-only: false
```

This causes the cache to get hit when it shouldn't. In this case,
causing a change in `rust-toolchain.toml` to get missed in the file
hash calculation.

SIgned-off-by: Michael Kubacki <michael.kubacki@microsoft.com>