Docs: Add Recommendation to Use Debugger Feature Flag
After analysis with cargo-bloat, it has been determined
that the debugger, even when disabled, adds 60k - 200k of
file size to a platform, depending on the platform. This
updates the patina docs to recommend a feature flag to
enable the debugger in order to save this space when it
is not in use.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>