patina_debugger: Remove initial breakpoint configurability.
The `configure` function in PatinaDebugger currently allows the platform
to dynamically enable the debugger without the initial breakpoint. Currently
this input is ignored as having the debugger enable in a silent configuration
can be potentially dangerous for having the debugger enabled without notice in
non-development scenarios. This removes this option and changes `configure` to
`enable` for clarity.