component params: Bugfix and additional testing
While writing tests in this same commit, a bug was found in the param
initialization / validation logic for ConfigMut<T>. The wrong panic
message was being displayed when there was a conflict in parameter
accesses for a param function. This commit fixes that bug.

This commit also increases test coverage for the param module.