patina: Add Arch Specific mod and Move [read|write]syreg there
There are many arch specific functions that need to be called
regularly and it would be valuable to have a home for them in
the patina SDK. This commit creates an arch specific module there
to host these.

The first use case is the [read|write]sysreg macros that are
currently duplicated in multiple places in patina, in different
forms. This unifies them in one location and expands them to
support more use cases.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>