patina_internal_cpu: Update Paging API to Always Return Cache Attrs
Currently patina_internal_cpu will not return MTRR cache attributes
if a page is unmapped; it will simply return an error. However,
the GCD code needs to send the cache attribute changed event to
MpDxe to wake up the APs and sync MTRRs whenever the cache attributes
change. The GCD took the approach of always sending the event when
a page was unmapped and is getting mapped; that resulted in large
numbers of the event getting fired, each having to wake up the APs.

This changes the patina_internal_cpu API to return the cache attributes
if covered by MTRRs even when the page is unmapped. For ARM64, where
the cache attributes are entirely contained within the page table and
the MP code starts cores fresh every time with the BSP page table, so
syncing is not required, cache attributes will never be returned if the
page is unmapped.

This greatly reduces the number of times the event is fired to only the
times it needs to.

This PR also updates to paging v10, which just included a breaking
change to update the API to not use a custom Result type.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>