dxe_core: Always Apply NX to UC Memory
This commit adds a new memory rules function that is easy to
audit the rules that Patina applies to memory. For now, this
function consists of a single rule, which is if EFI_MEMORY_UC is
set on a range, EFI_MEMORY_XP must also be set.

In DXE, we should never be executing from UC memory and on AArch64
it is architecturally defined as a programming error to have device
memory (what UC maps to) that is not marked XP. EDK II's ArmCpuDxe
always sets XP when UC is requested.

It has been observed that speculative execution on AArch64 platforms
can attempt to do an instruction fetch from device memory (which is
allowed if XP is not set) and crash the system if the region touched
is MMIO that does not have backing memory (say space between device
registers).

For all architectures, Patina enforces that UC memory should not be
executed from.

This commit also makes the XP requirement for all added memory
more explicit in adding resource descriptor HOBs, though this is not
a functional change.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>