Debugger: Support Reading MTRR/MAIR via Monitor Command
This commits an mtrr monitor cmd that accepts an address
and returns the cache type for that address.

This also updates the page fault exception handler to
print the caching type since that is not present in
the page table dump.

In order to allow uefiext to parse the cache attributes
in the page table walk, print MAIR_EL2 in the arch regs
monitor command.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>