Makefile.toml: Clean Coverage Artifacts Before Running Coverage
cargo-llvm-cov automatically cleans coverage artifacts before
running in order to have accurate coverage results. However, if
--no-report is passed, it does not automatically clean the coverage
artifacts.

Commit da004c7e5b942660c9b0b15b788af3a31ea18434 added --no-report
to cargo-llvm-cov when refactoring the report output, but did not
add a call to cargo llvm-cov clean first. This is fine for CI builds,
but for local runs, it will produce inaccurate coverage if there are
existing coverage artifacts.

This fixes that by ensuring cargo llvm-cov clean is called first.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>