GCD: Convert for_each_desc_in_range to an Iterator
For easier usability, for_each_desc_in_range is converted
to an Iterator. This iterator needs to lazily find descriptors
because it gets invoked before memory is fully populated, so
we cannot rely on a Vec and finding all descriptors greedily.

This updates all usage to the new iterator and adds an
impl MemorySpaceDescriptor to return the overlap between
a caller passed memory range and a memory descriptor.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>