patina_dxe_core: Create static reference to Self
This commit adds static `__SELF` pointer to enable accessing "Self" from
a static context for usage in "efiapi" compatible functions which cross
the FFI boundary and thus cannot capture state like lambdas.

This effecively makes the core a Singleton in terms of efiapi functions
and thus this change assumes that only a single instance of the Core
will be instantiated.

Please noate that the only functions allowed to use `Self::instance()`
are "efiapi" functions. No internal usage of `Self::isntance()` is
permitted`.# Please enter the commit message for your changes. Lines starting