patina_dxe_core: Tests: Fix Intermittent Test Failure
Recent tests added a MockPageTable to the static GCD, but
this is not cleaned up on GCD.reset(). As a result, there is
a race condition where certain tests that expect there not
to be a page table installed in the GCD can fail. This fixes
that by removing the page table if installed in GCD.reset().