Component state: Use Cow<'static, str> over &'static str
This commit updates the Component state use Cow<'static, str> rather
than directly a &'static str for component state (including the name
and param error message"). This greatly improves the API available to us
when working with these fields, while ensuring we only allocate a new
instance of the string if we absolutely have to. As a side note, no
allocation is made if creatining a Cow from an already static string, so
there is no negative impact to this change.