patina: `Param::init_state` to return Result
This commit updates the Param trait's `init_state` method to return an
result, where the error is a string explaining why the parameter could
not be initialized.

This error is ultimetly bubbled up to the to the component where the
component_dispatcher will place the component in a separate list of
components that failed initialization. The component is eventually logged
with the reasoning why it failed to initialize.