patina_dxe_core: store image_device_path in Box
This commit updates the `PrivateImageData` structure to store the
image_device_path protocol in an Optional Box, instead of a raw pointer
so that we can properly and automatically manage the lifetime of the
protocol without needing to leak it.

This commit also updates the `core_unload_image` logic to not drop the
`PrivateImageData` if it fails to uninstall the loaded_image /
loaded_image_device_path protocols or to remove the runtime image.